Begin the switch to em-proxy based solution. This probably breaks a lot.
This allows the proxy to also inspect the outgoing http responses so it
can gather more analytics.

This is a big code drop based on some old, uncommitted code. It's been
updated and reworked a bit, but it still needs some more love.
Functionality is now likely broken, and various things still need to be
cleaned up.