Switch to using nginx instead of haproxy as the main router.
This is due to HAProxy's current lack of server-side keep alive support. It's
coming in HAProxy 1.5, but the feature has yet to land, so at least for now
we're switching to nginx. HAProxy may have more robust load balancing and
proxying capabilities, but the server-side keep-alive feature actually improves
the performance quite a bit for our main use-case of remote backends. While it
doesn't make much of a difference if all your servers are local, since we're
routing to servers all over the place, being able to hold open a pool of
cached, persistent connections for a short time speeds things up a fair amount
(I've experienced improvements of 10-100ms per request).

We could revisit once this functionality lands in HAProxy 1.5, since HAProxy
still has a lot of powerful features that might be nice to take advantage of.
But nginx's config syntax makes some things a lot easier, so that's also a nice
bonus of the switch to take into account.