A messy, incomplete first stab at an admin UI for managing APIs.
Parts aren't complete yet, but the overall flow is starting to take shape for
how APIs could be managed online.

Given how nested these forms are, and the Uflow I'm thinking of, I'm
experimenting with using Ember.js to manage this part of the admin. If Ember
works out, it would probably make sense to standardize and switch the simpler
analytics UI from Marionette to Ember.