A messy playground for switching the admin stats over to Ember.js.
Since Ember seems to mostly be working out for the API admin, I'd like
to standardize on a single MVC framework for the admin javascript,
rather than also using Backbone/Marionette only for the stats.

So far all that's really implemented is a new approach to fetching the
data for the logs table and rendering using DataTables. However, this
fixes a number of pesky oversights in the previous implementation
(previously it was only returning the first 500 results client-side, so
sorting and pagination only sort of worked).