Shift rate limits to the settings config area.
This will allow for rate limits to be overridden on a per service basis
(like other settings). This also moves things closer towards allowing
per-user rate limits.

As part of this, IDs for tracking settings objects are now automatically
generated for the config items as necessary. This just makes it a little
friendlier to define these items in the YAML config or tests (when we're
not using the database with automatic ids).