Web admin UI for custom, per-user rate limits.
The per user rate limits are now the more granular type that can be
defined on any interval, like the new API backend limits.

The whole API user management interface has been Ember-ified so we can
leverage the same nested dynamic forms we use for managing these rate
limits for APIs.

This also fixes some bugs with the Selectize Ember view when the model
data isn't present on load, and is loading from ajax. In that case, the
Selectize view was being left empty. Now it stays in sync with the model
data even if it wasn't present on initial load.