Loosen DNS resolver so it caches and tries to reuse recent IPs.
As long as an IP has been seen in the past 4 hours, it will continue to
use it. This slightly breaks DNS TTL, but this was needed to accommodate
domains that have very short TTLs and cycle through a variety of IPs.
That was causing new IPs and nginx to be reloaded too frequently. 4
hours may be overkill, though, so this could be revisited if that turns
out to be too long.