Fix logging when errors are being returned.
When the Gatekeeper was responding directly with errors, the event that
gathered the analytics and logged them wasn't being triggered. Now all
errors that are handled by the errorHandler util (which should be
everything) fires a similar end event that can be captured so we can
deal with logging similarly.

Fixes NREL/api-umbrella#13.