Some ugly, temporary workarounds for DNS resolution issues.
Basically, all this temporarily works around issues with our DNS
resolver and TTLs. Our DNS resolver was semi-ignoring short TTLs, so we
didn't continuously restart nginx when dealing with hosts like Akamai,
where we seem to constantly get two different IPs back that are both
valid. However, our over-agressiveness with TTLs turned out to break
other hosts, like ELBs, where the IPs changed less frequently, but
when they did, it was important that it be reflected.

This isn't a great solution, but hopefully it will temporarily do the
trick. The real solution is on the config branch, where we're splitting
up the nginx servers, so then we can more readily restart the router
nginx server whenever we please (that should allow us to actually
respect TTLs, and clean all of this DNS business up a lot).