Enable gunzip module in nginx to work with backends that always gzip.
I'm not sure how important this behavior really is, but we were already
testing for it, but it turned out Varnish is what was providing this
behavior for. It didn't seem super clear in Varnish's docs that it would
actually support this behavior, so shift this behavior into explicitly
into nginx. This is so we can experiment with other HTTP caching layers
that doesn't supply this functionality (Traffic Server).