adding link to main image
When on http://nrel.github.io/api-umbrella/, I spend a decent bit of time unsure what I can click.  It looks to me like only the tabs at the top are clickable.  I recommend making the entire image in the body section a link to the other page so that users can more easily move forward.