Shift stripping analytics cookies for caching into the gatekeeper.
This is partially done so we can experiment with apache trafficserver as
the caching layer (where this type of replacement on the cookies is more
difficult than in Varnish's VCL). But I think this approach also makes
it easier to configure (since we can just focus on the cookie names),
and we could potentially make this more configurable down the road (for
example, different API backends may want to filter cookies differently).