Fix nginx reloading being broken in the config reloader.
This also fixes running the test suite. It turns out Nginx reloading had
been broken for a while, but 8d7ae67260902af093f327e9992cfac1d2644942
just broke the test suite due to changes in the order of the config file
writing (nginx now needs to be reloaded once during the startup process
to read in the initial config).