Setup to allow the default install to respond to any host.
Now an out of the box install will by default respond to any hostname
(so it's only with additional configuration that the hosts become more
specific).

Related to this, the app now generates a self-signed SSL certificate for
hosting purposes if one is not supplied. This is not what a real
environment would use, but useful for initial development environments
(without having to completely disable SSL there).