Remove environment name from database names.
This will make setup a little easier, and makes the app_env a little
less relevant. This partially stems from wanting to run the default
omnibus install in the "production" environment, regardless of what
purpose the install really serves (and this is so we always use
precompiled assets in the default omnibus install).