Enforce connection timeouts during connection test phase of startup.
Without this, elasticsearch was sometimes accepting connections during
startup, but then not serving up a response. This was leading to startup
failures, since the connection was never retried. So switch to the
"request" module for testing http connections, since it makes it much
simpler to account for timeouts or other types of errors.