Fix so wildcard host apis still match when a hostname match is present
Previously, the internal wildcard matching for the API Umbrella APIs
was only matching as long as an api didn't exist for the hostname being
used. But as soon as an api was added with the hostname being used, then
all the internal API Umbrella APIs broke, since they were using the
wildcard matching.