Change all the default internal ports to be in the 14000-14060 range
Previously we were using ports in the 50000+ ephemeral range. I'm not
sure why I ever thought it would be good idea to use ephemeral ports
(maybe to reduce chances of conflicting with existing server ports), but
since we hardcode all our ports (and don't retry on port conflicts), we
should not use ports in the ephemeral range.

Generally, this wasn't an issue, but I just ran into a port conflict for
the first time due to some other service deciding to take one of these
ports on boot.

This also adjusts how the gatekeeper ports ranges are defined, so
they're not hardcoded in the supervisor file. This consolidates all the
port configuration to the one yaml file.

This also adds some missing port definitions to the test.yml file, so
running tests shouldn't conflict with a server that's also running a
real instance of api umbrella.