Fix missing data from router.log for rewritten requests.
Requests with this /api prefix were being properly handled, but then the
`$x_api_umbrella_request_id` and `$router_name` variables were missing
out of our log files. This is apparently due to the ordering of the
rewrite and its use of `break`. Odd. But shifting the rewrite later
fixes things.

More generally, we probably shouldn't have this /api rewrite in the
global config like this. We should definitely remove this from the
public repo at some point, and maybe we can even remove it altogether
for api.data.gov's use too.