Setup /metrics/ so it gets routed to the static website content.
This is specific to the api.data.gov site right now, so this routing
setup isn't very ideal. What I'd really like to do is switch things
around so API backends are what get white-listed for routing, and the
website receives other unknown requests, but that requires a bit more
work and planning.