Attempt to cleanup keepalive server tests to make more reliable.
The main confusion was that the keepalive number is per nginx worker
process, not global. We've also increased the concurrency to try to
saturate the connections more reliably
(9799f15d1d0495602897e5069a1f41aabf8a01f5 should also help with this to
ensure we're not hitting # of threads/processes limits internally and
lowering the concurrency).