Begin refactoring log collection process.
The primary change is that the frontmost nginx logs become the primary
source of data, so we're only gathering some timing information from the
gatekeeper level we're interested in logging.

The complex redis queuing is also going away in favor of using
beanstalkd for log queue (but redis is still being used to coalesce the
logs data from differnet sources).

Finally one new addition is logging the "deny code" when the gatekeeper
blocks a request. This will make it much easier to view in the analytics
which requests never made it to the API backend and for what reason (eg,
api key not passed in, api key not valid, over rate limit, etc).