Revamp logging to use Bunyan and cleanup how Rollbar errors are logged.
I'm not completely wild about Bunyan's default JSON log output, but
their API makes it a lot cleaner to log metadata associated with log
entries, which in turn makes it a lot easier to send that metadata to
Rollbar in a structured way when we have errors. Bunyan's API also makes
it much easier to log to multiple sources (like only to Rollbar for
warnings and above). All in all, this cleans up a lot of things that
were beginning to feel kludgy with our own custom logging
implementation.