Some very initial experiments with overhauling the dns resolution stuff.
I guess this is mostly commented out in this commit, but wanted to make
note of the possibility to revisit some of this DNS business using the
new jdomain plugin for nginx:
http://wiki.nginx.org/HttpUpstreamJdomainModule This could do away with
the need for our custom DNS resolver, and allow us to more efficiently
handle all of this directly inside nginx. This would cut down on nginx
reloads and allow us to truly respect all domain TTLs. Fingers crossed.
As part of this, also experimenting with running a local dnsmasq server
for DNS caching purposes.