Fix analytics results for limited admins with uppercase scope paths.
If an admin belonged to a scope that had mixed cases letters in the path
prefix, the analytics for those scopes wouldn't show up in the analytics
section. This was being caused since the elasticsearch index for the
"request_path" field uses a lowercase analyzer.

We really need to add a test suite around this analytics stuff in the
admin, but we first have to setup a test elasticsearch database
connection, figure out how we want to populate test elasticsearch data,
ensure we're using the template we use on our real servers (so things
like our lowercase analyzer for certain fields would be caught), etc.