Rollbar has to be required earlier than the initializer to work.
I think something may have changed in the Rollbar gem, since our
optional require in the initializer used to work with rollbar 1.1.0. In
any case, Rollbar is still optional, it's just that it now gets required
in earlier. But if you don't specify a key, then Rollbar shouldn't do
anything.