Fix analytics CSVs. Disambiguate analytics start/end time params.
See: https://github.com/18F/api.data.gov/issues/173

When we updated DataTables to 1.10 and the new ajax standard
(http://datatables.net/manual/server-side), we began to have conflicts
over their concept of the "start" param being for pagination, and our
use of "start" parameter being for date ranges. We had sort of solved it
by renaming the parameter differently depending on the use-case, but
this got to be confusing. This is also what partially broke the CSV
downloads, since the query parameter names had gotten out of sync.

This fixes all of that by letting DataTables take the "start" parameter,
and we've renamed our date parameters to be "start_at" and "end_at".
This removes the conflict and makes things a little more sane.

There were other aspects of the CSV download that were also broken from
when we switched to the official elasticsearch ruby client. Those are
also fixed now.

And finally, we've finally begun to add some very initial tests to the
analytics area, so we're actually testing our elasticsearch queries.
Tests are now present for the CSV download link to hopefully make sure
this doesn't happen again.