Fix router reloads/deploys not picking up latest code changes.
Since we were relying on the PATH set in lib/process_env.js to pick up
the binary paths, this PATH was only being set at start-time and not on
reloads. Since the PATH being set contained a resolved path, it meant
that the "current" symlink used for deployments got resolved at start
time, so the processes would forever use that deployment folder (eg,
router/releases/20150126015553 instead of router/current). This
prevented reloads from ever actually picking up code changes to the
router project.

This fixes it by using the "current" path in our paths to the binaries,
so the symlink will be resolved each time. This mirrors how we handle
the current path for the web project (and why the reloads were working
for that project).