Extend timeout for analytics tests to fix periodic failures.
This timeout shouldn't need to be this long, but for some reason, the
"api_backend_router" logs sometime take an extra 5-6 seconds to show up.
This doesn't happen when the logging tests are run by themselves, but it
does happen consistently when the entire test suite is run. The timers
on the "api_backend_router" are also affected (consistently 5 seconds
longer than they should be). I think it's somehow related to keepalive
connections, but since the front-most timers don't reflect this longer
response time, I think it's okay to ignore.