Fix url escaping issues around the "route" type of rewrites.
This also adds a lot more tests surrounding the route pattern type of
rewrites (and these tests are also what uncovered the URL encoding
issues). Most of the route pattern testing was being covered by our use
of the https://github.com/bjoerge/route-pattern library, but this
adds more concrete tests on our end, so we're actually testing the
expected behavior from our stack with the use of that library. This main
stems from the Lua branch experiments and wanting this functionality
tested, so we can reimplement it in a compatible way on the lua branch.