Add a more specific test for rolling reset behavior of rate limiting.
This test ensures that no more than the hourly limit can be called
within the past hour. This is achieved by our bucketing approach to rate
limits, so we can always know how many requests were made in the past 60
minutes.

This differs from a simpler approach that I had initially implemented on
the Lua branch, so I wanted to to more explicitly test our current
behavior. The simpler approach is based purely on a single hourly bucket
with a 1 hour TTL. While simpler, that approach allows for nearly double
the hourly limit in this scenario:

- 8:14 - 1 request made
- 9:13 - 999 requests made
- 9:14 - 1000 requests made

Since the 9:13 data would have gone gone into the bucket created at
8:14 and then immediately expired, since that bucket had a TTL of 1
hour, this then allows nearly double the hourly rate limit within 2
minutes.

Our currently implemented approach is a little heavier, but avoids this
issue by bucketing the request on a per minute (or whatever accuracy is
specified) basis, and then summing each of those buckets that exist
within the past hour.