Re-implement rate limiting to bring it inline with current behavior.
Rate limiting now resets on a more granular rolling basis. See this
comment for why this change is necessary versus the simpler
implementation we had made on the Lua branch initially:
https://github.com/NREL/api-umbrella-gatekeeper/commit/c289f73d49c8aab622a6fa2c6a200d6c1a42e5d8

The distributed pusher and puller still need to be updated for this new
algorithm, but this gets most of the rate limiting tests functional on
this branch.