Fix tests and introduce messy first stab at DNS resolving.
This gets the tests largely back to where they were prior to the router
integration merging on a87d1fb76d6c8dcc31f207c4041bbb32599fb438. That
is, most of the gatekeeper tests pass, but not quite everything (about
9 failing and 386 passing tests).

Part of the test failures were due to long the dyups module blocking
when resolving non-existent domain names. So this is a rudimentary first
stab at shifting to a non-blocking resty-dns resolver.