Initial query builder interface for analytics queries.
This uses the jQuery-QueryBuilder plugin to provide an easier
query building interface for constructing analytics queries. This is
intended to finally provide a nicer alternative to the raw "Advanced
filters" query syntax, which is just a straight ElasticSearch query.

I still need to figure out how the interface for this will fit in, but
it basically seems to work.