Experimenting with Ledge for HTTP caching layer.
Unfortunately, this seems to break a lot of our expected proxy behavior
(since it's really bypassing nginx's default proxying capabilities and
proxying with Lua HTTP libraries instead), so I'm not sure this will pan
out without more work.

The other alternative is nginx's built-in caching, which seems simplest
to use and maintains most of our expected proxy behavior. However, it
seems to not work at all with buffering disabled, which is pretty
important for our proxy. See: http://serverfault.com/a/258225