Experiment with TrafficServer as caching layer.
Since it seems like we might go back to a separate caching server
(rather than the initial thought in the openresty work to use nginx's
built in cache or Ledge), then this continues the experiments with using
TrafficServer instead of Varnish. It seems to benchmark a bit better,
and there are some other nice things about TrafficServer. It also has
built-in DNS handling for dynamic backends, so we may also experiment
with using TrafficServer directly for the API backends proxy, rather
than going back to nginx (since having TrafficServer handle all the DNS
stuff would certainly be simpler).