Fix conflicting model in tests with the api-umbrella-config module.
I'm not sure why this didn't happen before, but I think it might be
related to getting the api-umbrella-config module bumped and on the same
version of mongoose as the router and gatekeeper.

This could be handled better, but in the meantime, we'll just call the
router one something different.