Fix unlimited rate limit on APIs not working for users with settings.
An API with an unlimited rate limit mode was not being applied when a
user had been edited in the admin and had a default rate limit selected.
In this case, the "null" value for the rate_limit_mode at the user-level
was overwriting the "unlimited" value that should have been taking
precedent at the API level.