Fix broken redirects for the web admin login due to routing changes.
In the vagrant environment, trying to request https://10.10.33.2/admin/
ended up redirecting you to https://localhost/admin/login. So we need to
go back to preserving the host headers we pass to the web app and during
redirects. Now we only normalize the host header for the APIs, where we
can more properly use the normalized header for matching purposes.