Fix logging of request scheme (http vs https) when behind load balancer.
If API Umbrella was sitting behind an SSL terminator (like an ELB), then
we weren't properly logging the original protocol of the request from
the X-Forwarded-Proto header.

This has been fixed and tests surrounding this, as well as a sanity
check on the logging of the Host header have been added.