Remove legacy "/api" prefix stripping, since that's no longer effective.
Due to the recent routing changes revolving around APIs and website
routing, this /api prefix stripping was no longer in use (since the /api
prefixes would never make it to the gatekeeper to be stripped). Since it
doesn't look like anyone was really using the prefix in production,
let's clean it up, since it seems like it might be something that could
eventually bite us (if we do want a /api route).