Return X-Cache response header. Use to fix unexpected logging warnings.
This X-Cache header makes it much easier to distinguish between cached
responses and uncached responses (instead of having to rely on the Age
header).

This allows us to fix some periodic log warnings indicating that the log
data was incomplete, when it actually wasn't. This would occur when
simultaneous requests would come in and the first one got cached. In
this case, we don't expect log data from the "api_backend_router", but
our previous reliance on the Age header to detect cached responses made
this difficult to handle, since the newly cached responses would all
have an age of 0.

We're also now escalating the logger level when there is unexpected
missing log data if the response was an error. This should hopefully
help better handle situations like the rare 503s Varnish was generating
(see 2e9007c983e930a13db5df6510249ad908634398) so we can be better
notified.