Fix logging failures when query string field first indexed as date.
If a request came in with a query string like `?something=2015-01-01`,
elasticsearch was automatically indexing the `request_query.something`
field as a date. This would work fine if all the subsequent requests
contained valid date strings for that "something" field. However, if
another request came in for `?something=foo`, the indexing for that
request would fail. This was because elasticsearch had already mapped
`request_query.something` as a date, and "foo" failed to parse as a
valid date.

Now we treat all query string parameters as just strings in
elasticsearch, since we can't really predict types on these fields, and
we don't really need type information on these `request_query.*` fields.