Fix reload temporarily resulting in 404 errors. Add tests around reload.
Due to the website routing changes introduced in
https://github.com/NREL/api-umbrella-router/pull/6, the reload command
was now resulting in a temporary 404, due to the new frontend_hosts.conf
being written too early during the reload process. Now we defer writing
it like how we handle the backends.conf file.

This also adds integration tests around the reload feature to make sure
it always results in a zero downtime reload and to help prevent this
type of regression in the future.