Fix XSS escaping issues in the admin data tables.
I mistakenly thought DataTables escaped things by default, but this is
not the case. Escape all data tables fields now and add capybara
integrations tests to ensure this is effective.

Setting up the Capybara tests required a slight change in the analytics
"Filter Logs" table, since the ajax request needs to be performed via
POST instead of GET. I think this is due to the length of the URL
involved in the query. This had worked in most browsers, but I'm
guessing this was broken in IE anyway, so it's probably a good idea
anyway.