Fix intermittent integration test failures. Upgrade gems.
This mainly fixes an occasional test failure relate to the download CSV
integration test failing.

I think the gem updates may also fix a rarer test failure related to the
datatable's spinner displaying. So the gem updates are more a side
effect of debugging this, but they seem to be working fine.