Try to prevent duplicate config reloads. Better config change logging.
It was possible that multiple config reloads would be triggered if there
were both file changes and DNS changes detected at the same time.

Try to improve the logging a bit during config reloads. Add a diff-based
output of the config files being written when there are changes. This
seems like it might be the most terse, yet understandable way to debug
and decipher what config changes are actually being made to trigger
reloads. As part of this diff-based output, tighten up the handlebars
templates to get rid of extraneous white space so that the diffs are
more relevant.