Switch to mora (go backend) for  mongo queries to support replicasets.
While it would be nice to use a native mongo library for openresty, the
current Mongo support for OpenResty scene isn't great, and adding proper
support for things like replicasets seems a bit more than I want to try
to implement myself right now in those adapters. So instead, we'll use
mora, which provides a lightweight server process and REST API on top of
MongoDB. Performance actually seems to be pretty close to what we
experienced with the native openresty mongol adapter, and this approach
(treating the database as an API), actually isn't too far removed from
how openresty has historically accessed databases (like postgres), so
while not ideal, it doesn't seem terribly horrible in the OpenResty land.