Initial experiments with using shcache for loading API user records.
There are a couple different reasons behind this move versus the current
approach where we replicate the entire user database locally in an nginx
shared dict:

- This new approach allows for the user database size to exceed the
  nginx shared dict size. Old items would be pruned from the shared
  dict, but then fetched again when needed (rather than needing to be
  careful about the shared dict size and ensuring it's big enough to
  store all users).
- This reduces the hyper-frequent polling of MongoDB looking for user
  changes. And related to that, since we're no longer reliant on
  polling, there's no no longer any edge cases with a new user
  making a request before their user account data was fetched by the
  poller.

The one thing we haven't yet solved is how to invalidate these cached
records. We either need to introduce relatively short TTLs on the cache,
and then just deal with whatever delay that may cause for updates. Or
alternatively, we could still poll MongoDB to detect changes, but at a
reduced frequency (maybe every second), and invalidate the cache that
way.