Experimenting with a makefile to install all our bundled dependencies.
This is a replacement for the Omnibus-based package installers we've
setup. We'd still want to make packages out of these Makefile-based
installations, but I think this approach might be much more
straightforward.

It allows us to upgrade individual packages with a bit more ease, rather
than having to go through the entire build process separately. This
would also make it easier to tie code changes with dependency changes,
since they'd exist in the same repo (this hit us recently when we wanted
to add tests in the router project, but they wouldn't actually pass
until new packages were fully built and released, all of which made
testing a bit difficult).