Fix formatted_errors tests for lua branch.
This does change the behavior of a couple of the format corner cases,
but we were more concerned with making sure these cases don't die,
rather than trying to treat them in any specific way, so I think this is
okay. The changes are:

- "format[]=xml" no longer gets treated as XML. Instead, it's treated as
  unknown (falling back to JSON).
- "format=xml&format=csv" now parses out the first match, so it's
  treated as XML, rather than unknown.

Both of these are due to differences in how nginx parses query
parameters, but I think we're okay with this behavior, and there's no
need to replicate the old behavior for these corner cases.