Rework log indexing to use ElasticSearch bulk indexing.
This allows the elasticsearch indexing to happen much more quickly and
more efficiently.

I'd still like to revisit a lot of this logging stuff, but for now, this
seems like the quickest way to improve the logging so it can recover
more quickly from a backlog, while still keeping our basic logging
the same.