Add ability to have wildcard backend hostnames.
This extends the wildcard hostname support added in
https://github.com/NREL/api-umbrella-gatekeeper/pull/13 Now the backend
host can be rewritten to also contain the matched piece of the
wildcard in the frontend host (previously the backend host was either
static or an empty string, in which case the frontend host was passed
through verbatim).

The backend host just uses the same wildcard `*` syntax for the
replacement value. Now with a frontend host like `*.example.com` and a
backend host of `*.api-backend.foo`, a request to `penguin.example.com`
would be rewritten into `penguin.api-backend.foo`.