Implement different approach to API reordering to reduce reorder churn.
This reduces the need to publish multiple API Backends when only
reordering a single API backend for matching purposes. This becomes more
important when there are admin groups and scopes in use, since it
prevents one admin group from affecting another group's APIs when
performing a reorder.

Previously, we ordered each API backend with a numeric counter like 1,
2, 3, 4, 5, etc. So if you wanted to move the last backend to the front,
you actually ended up having to change the order for all the backends.

The new approach is similar in that it uses a counter, but we provide a
significant gap between each item (10,000). So instead, API backends
added would by default have a sort order of 0, 10000, 20000, 30000,
40000, etc. If you wanted to shift the last backend to the front, it
would get a new sort order of -10000, while leaving the rest of the
backends untouched. More important than simply going negative, if you
wanted to move the last backend to the second position, it would leave
the rest of the items untouched by choosing a sort order in between the
existing values (so 5,000 in this example). Subsequent moves in between
the first two items would continue dividing the available integer range
in half (2,500, 1,250, etc.)

With this approach of leaving gaps and then filling them in to fulfill
moves, there is a theoretical point where the surrounding items will
need to be shifted. Similarly, the sort order uses 32bit signed
integers, so there's also a theoretical maximum upper bound and lower
bound if enough entries get added with the gaps in place. Both
situations should be handled in this implementation by shifting
surrounding records to make room as need. In those cases, other records
may still be changed when sorting, so this isn't a 100% perfect
solution, but I think both situations should be exceedingly rare given
the amount of data typically present and number of reorders that
typically happen.