Add option to override roles in sub-settings. Some refactoring...
The override roles checkbox is part of some changes to how roles get
applied and inherited made in:
https://github.com/NREL/api-umbrella-gatekeeper/commit/94670b0097f044a4017a3afa1ad25dc7c2171280

This turned into a larger change than just adding the single override
checkbox for roles, since we only want this checkbox to show up in the
context of the sub-url settings, and not the other places that those
settings fields are shared (api global settings). So we can more easily
control pieces of the settings fields, it's been turned into an Ember
component, instead of a render call, which gives us the ability to pass
extra variables around. This in turn led to a few other things:

- Turn the selectize input from a sort of strange view wrapper into an
  easyform input type. This should be more straightforward, and since
  Views will be removed in Ember 2, it seems like a good move to go
  ahead and make.
- Tweaks to the selectize and ace editor inputs to improve usability and
  testability. Now the associated labels for these input types are
  correct, so clicking on the labels lands you in the correct input
  field, rather than still being tied to the underlying, hidden field.
  This and some related changes help with testability of these input
  types.
- More shifting of text in the api and settings forms to i18n while
  we're at it.
- A much more complete full integration test for the whole API backend
  and API user sections. Since we needed to shift the shared settings
  from a view/controller to a component, we needed a much more thorough
  integration suite to ensure this didn't break any piece unexpectedly.