Disable some integration tests that need to be rewritten.
There are still other failures with this lua rewrite, but these tests in
particular cause the entire test suite to bomb. So let's remove them for
now so we can at least get a better sense of where the other tests are
in terms of pass/fail for the lua branch.